+++ CHANGELOG.md
@@ -1,6 +1,6 @@
+- Use plain TextDecoder for decoding autoSql
+
 ## [2.0.3](https://github.com/GMOD/bbi-js/compare/v2.0.2...v2.0.3) (2022-10-10)
-
-
 
 - Add BigInt64 polyfill for older safari